Don't split player arcs
Instead, to avoid collision with previous segment, keep track of absolute amount of turn on each arc, we need to have "travelled" at least 1 circle before a collision with own trail.
This should stop the bug where it was possible to pass through an arc at the point where either arc was split (note this is still possible if the player initiates a direction change at this point, but hey lets call that a "feature"!).
Tests not updated to pass.